Add test scripts and test files
The actual Playwright test can be found in PublisherInterface.spec.ts.
This file should not be modified.

The tests ran against Publisher are in the ./test/tests folder. You can
add your own tests but please read the README.md as it contains useful
information for writing tests.

The main script which does all the magic is run_tests.js. This script
builds, copies, starts the server, and runs Playwright. It was
necessary to use a script due to the complexity of testing Publisher
which was not designed for testing. Read the README.md for more
information.