[FEATURE] Tests (#13)
* Add playwright as a dev dependance with command

Playwright is used to make testing with Publisher in an live iframe
possible.

* Add test scripts and test files

The actual Playwright test can be found in PublisherInterface.spec.ts.
This file should not be modified.

The tests ran against Publisher are in the ./test/tests folder. You can
add your own tests but please read the README.md as it contains useful
information for writing tests.

The main script which does all the magic is run_tests.js. This script
builds, copies, starts the server, and runs Playwright. It was
necessary to use a script due to the complexity of testing Publisher
which was not designed for testing. Read the README.md for more
information.

* Update README

* Update testServer to resolve issues on Linux

* Add more Linux fixes to test scripts

The killAllTestServerProcess() is a hack around issue had on
Ubuntu 22.04 with spawn with shell:true creating 2 processes for
every one command.

* Modify tests to use createInterface function

This is to allow options to be tested for future functionality. Without
this ability, it would be impossible to test options passed to the build
method.