Add alias for editorObject
Many integrations will need to move over to Publisher Interface and all
there code will use PascalCase function calls. This provides an easy
alias to make it easier to those code bases to upgrade to Publisher
Interface.

This does not solve the fact that these functions still return a
Promise.