Update Do I Need To Convert My Integration
Reordered the Does this effect me and the Why? The table looks nice
but with it in the suggested order, it misses context. However, now
it is setup so that one can easily see the table, and skip the why
context if they choose.

Removed <a> tags, as it is not the style throughout other markdown.